Object class to separate file
